requirements.txt
pillow
scikit-learn
opencv-python
numpy
scipy
PyMaxFlow